Separate tests for clarity and consistency with #832
